Store active bindings for cloning
In R-devel, as.list.environment no longer returns the function for an active
binding; instead it returns the value from executing the function. When the
object is cloned, we need to get the function definition, but there's now no
way to get it from the environment, so we have to store a copy of it.